+++ compiler/core/lam_analysis.ml
@@ -46,8 +46,8 @@ let rec no_side_effects (lam : Lam.t) : bool =
       | _ -> false)
     | Pcreate_extension _ | Ptypeof | Pis_null | Pis_not_none | Psome
     | Psome_not_nest | Pis_undefined | Pis_null_undefined | Pnull_to_opt
-    | Pundefined_to_opt | Pnull_undefined_to_opt | Pjs_fn_make _
-    | Pjs_fn_make_unit | Pjs_object_create _ | Pimport
+    | Pnull_undefined_to_opt | Pjs_fn_make _ | Pjs_fn_make_unit
+    | Pjs_object_create _ | Pimport
     (* TODO: check *)
     | Pmakeblock _
     (* whether it's mutable or not *)