+++ compiler/ml/translcore.ml
@@ -372,7 +372,6 @@ let primitives_table =
       ("%null", Pnull);
       ("%undefined", Pundefined);
       ("%is_nullable", Pisnullable);
-      ("%undefined_to_opt", Pundefined_to_opt);
       ("%null_to_opt", Pnull_to_opt);
       ("%nullable_to_opt", Pnullable_to_opt);
       ("%function_arity", Pfn_arity);
@@ -397,7 +396,6 @@ let primitives_table =
       ("#is_nullable", Pisnullable);
       ("#null_to_opt", Pnull_to_opt);
       ("#nullable_to_opt", Pnullable_to_opt);
-      ("#undefined_to_opt", Pundefined_to_opt);
       ("#makemutablelist", Pmakelist Mutable);
       ("#import", Pimport);
       (* FIXME: Deprecated *)