+++ tests/tests/src/js_undefined_test.mjs
@@ -2,14 +2,13 @@
 
 import * as Mt from "./mt.mjs";
 import * as Js_undefined from "rescript/lib/es6/Js_undefined.js";
-import * as Primitive_option from "rescript/lib/es6/Primitive_option.js";
 
 let suites_0 = [
   "toOption - empty",
   param => ({
     TAG: "Eq",
     _0: undefined,
-    _1: undefined
+    _1: Js_undefined.toOption(undefined)
   })
 ];
 
@@ -19,7 +18,7 @@ let suites_1 = {
     param => ({
       TAG: "Eq",
       _0: undefined,
-      _1: undefined
+      _1: Js_undefined.toOption()
     })
   ],
   tl: {
@@ -28,7 +27,7 @@ let suites_1 = {
       param => ({
         TAG: "Eq",
         _0: "something",
-        _1: Primitive_option.fromUndefined("something")
+        _1: Js_undefined.toOption("something")
       })
     ],
     tl: {